script modification
